+++ Sources/MapboxCoreNavigation/CoreConstants.swift
@@ -5,6 +5,8 @@ import MapboxDirections
 // MARK: RouteController Rerouting logic
 /**
  Maximum number of meters the user can travel away from step before `RouteControllerShouldReroute` is emitted.
+ 
+ - important: Only works for `LegacyRouteController`.
  */
 public var RouteControllerMaximumDistanceBeforeRecalculating: CLLocationDistance = 50