+++ Sources/MapboxCoreNavigation/NativeHandlersFactory.swift
@@ -20,6 +20,7 @@ class NativeHandlersFactory {
     let targetVersion: String?
     let configFactoryType: ConfigFactory.Type
     let datasetProfileIdentifier: ProfileIdentifier
+    let navigatorRouterType: MapboxNavigationNative.RouterType?
     
     init(tileStorePath: String,
          credentials: Credentials,
@@ -27,7 +28,8 @@ class NativeHandlersFactory {
          historyDirectoryURL: URL? = nil,
          targetVersion: String? = nil,
          configFactoryType: ConfigFactory.Type = ConfigFactory.self,
-         datasetProfileIdentifier: ProfileIdentifier = ProfileIdentifier.automobile) {
+         datasetProfileIdentifier: ProfileIdentifier = ProfileIdentifier.automobile,
+         navigatorRouterType: MapboxNavigationNative.RouterType? = nil) {
         self.tileStorePath = tileStorePath
         self.credentials = credentials
         self.tilesVersion = tilesVersion
@@ -35,6 +37,7 @@ class NativeHandlersFactory {
         self.targetVersion = targetVersion
         self.configFactoryType = configFactoryType
         self.datasetProfileIdentifier = datasetProfileIdentifier
+        self.navigatorRouterType = navigatorRouterType
     }
     
     // MARK: - Native Handlers
@@ -49,10 +52,16 @@ class NativeHandlersFactory {
     
     lazy var navigator: MapboxNavigationNative.Navigator = {
         onMainQueueSync { // Make sure that Navigator pick ups Main Thread RunLoop.
-            MapboxNavigationNative.Navigator(config: configHandle,
-                                             cache: cacheHandle,
-                                             historyRecorder: historyRecorder,
-                                             router: nil)
+            let router = navigatorRouterType.map {
+                MapboxNavigationNative.RouterFactory.build(for: $0,
+                                                           cache: cacheHandle,
+                                                           config: configHandle,
+                                                           historyRecorder: historyRecorder)
+            }
+            return MapboxNavigationNative.Navigator(config: configHandle,
+                                                    cache: cacheHandle,
+                                                    historyRecorder: historyRecorder,
+                                                    router: router)
         }
     }()
     
@@ -96,8 +105,26 @@ class NativeHandlersFactory {
                     endpointConfig: endpointConfig)
     }()
     
+    lazy var navigatorConfig: NavigatorConfig = {
+        return NavigatorConfig(voiceInstructionThreshold: nil,
+                               electronicHorizonOptions: nil,
+                               polling: nil,
+                               incidentsOptions: nil,
+                               noSignalSimulationEnabled: nil,
+                               avoidManeuverSeconds: NSNumber(value: RerouteController.DefaultManeuverAvoidanceRadius),
+                               useSensors: false)
+    }()
+
     lazy var configHandle: ConfigHandle = {
-        let customConfig = UserDefaults.standard.dictionary(forKey: customConfigKey) ?? [:]        
+        let historyAutorecordingConfig = [
+            customConfigFeaturesKey: [
+                "useInternalReroute": true
+            ]
+        ]
+        
+        var customConfig = UserDefaults.standard.dictionary(forKey: customConfigKey) ?? [:]
+        customConfig.deepMerge(with: historyAutorecordingConfig, uniquingKeysWith: { first, _ in first })
+        
         let customConfigJSON: String
         if let jsonDataConfig = try? JSONSerialization.data(withJSONObject: customConfig, options: []),
            let encodedConfig = String(data: jsonDataConfig, encoding: .utf8) {
@@ -106,14 +133,6 @@ class NativeHandlersFactory {
             assertionFailure("Custom config can not be serialized")
             customConfigJSON = ""
         }
-        
-        let navigatorConfig = NavigatorConfig(voiceInstructionThreshold: nil,
-                                              electronicHorizonOptions: nil,
-                                              polling: nil,
-                                              incidentsOptions: nil,
-                                              noSignalSimulationEnabled: nil,
-                                              avoidManeuverSeconds: nil,
-                                              useSensors: nil)
         
         return configFactoryType.build(for: settingsProfile,
                                        config: navigatorConfig,